PaltaTest/requirements.txt
aioredis==1.3.1
psycopg2==2.8.6
anyjson==0.3.3
asgiref==3.4.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
billiard==3.6.4.0
celery==5.1.2
certifi==2021.5.30
cffi==1.14.6
channels==3.0.3
channels-redis==3.3.0
dadata==20.7.0
python-decouple==3.4
Django==3.2.5
django-cors-headers==3.7.0
djangorestframework==3.12.4
django-celery-beat==2.2.1
httpx==0.18.2
humanize==3.10.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
msgpack==1.0.2
prometheus-client==0.11.0
prompt-toolkit==3.0.19
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==20.0.1
pytz==2021.1
redis==3.5.3
requests==2.25.1
rfc3986==1.5.0
service-identity==21.1.0
six==1.16.0
sniffio==1.2.0
sqlparse==0.4.1
tornado==6.1
Twisted==21.2.0
txaio==21.2.1
urllib3==1.26.6
uvicorn==0.14.0
websockets==9.1
zope.interface==5.4.0